tests/wpt/tests/infrastructure/expected-fail/unhandled-rejection-single-test.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Unhandled rejection in single-page test</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
setup({ single_test: true });
Promise.reject(new Error("error outside any setup or test"));
</script>
