LayoutTests/imported/w3c/web-platform-tests/infrastructure/expected-fail/unhandled-rejection-single-test-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: error outside any setup or test

FAIL Unhandled rejection in single-page test Unhandled rejection: error outside any setup or test
